backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.1
cron-descriptor==1.4.0
cycler==0.11.0
Django==3.2
django-celery-beat==2.5.0
django-extensions==3.2.3
django-filter==23.2
django-rest-framework==0.1.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
fabio==2023.6.0
flake8==6.1.0
fonttools==4.42.1
gunicorn==21.2.0
h5py==3.9.0
importlib-resources==6.0.1
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kombu==5.3.1
matplotlib==3.8.0
mccabe==0.7.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==3.5.3
referencing==0.30.2
rpds-py==0.10.3
silx==1.1.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
zipp==3.17.0
